Staging: brcm80211: remove uintptr typedef usage
Use a unsigned long, that is the proper type for this type of thing on
Linux.

Cc: Brett Rudley <brudley@broadcom.com>
Cc: Henry Ptasinski <henryp@broadcom.com>
Cc: Nohee Ko <noheek@broadcom.com>

+++ drivers/staging/brcm80211/brcmfmac/dhd_common.c
@@ -1477,7 +1477,7 @@ int dhd_iscan_print_cache(iscan_buf_t *iscan_skip)
 				   bi->BSSID.octet[2], bi->BSSID.octet[3],
 				   bi->BSSID.octet[4], bi->BSSID.octet[5]));
 
-			bi = (wl_bss_info_t *) ((uintptr) bi +
+			bi = (wl_bss_info_t *)((unsigned long)bi +
 						dtoh32(bi->length));
 		}
 		iscan_cur = iscan_cur->next;
@@ -1542,7 +1542,7 @@ int dhd_iscan_delete_bss(void *dhdp, void *addr, iscan_buf_t *iscan_skip)
 					bi->BSSID.octet[5]));
 
 					bi_new = bi;
-					bi = (wl_bss_info_t *) ((uintptr) bi +
+					bi = (wl_bss_info_t *)((unsigned long)bi +
 								dtoh32
 								(bi->length));
 /*
@@ -1568,17 +1568,14 @@ int dhd_iscan_delete_bss(void *dhdp, void *addr, iscan_buf_t *iscan_skip)
 							bi->BSSID.octet[5]));
 
 							bi_next =
-							    (wl_bss_info_t
-							     *) ((uintptr) bi +
+							    (wl_bss_info_t *)((unsigned long)bi +
 								 dtoh32
 								 (bi->length));
 							bcopy(bi, bi_new,
 							      dtoh32
 							      (bi->length));
 							bi_new =
-							    (wl_bss_info_t
-							     *) ((uintptr)
-								 bi_new +
+							    (wl_bss_info_t *)((unsigned long)bi_new +
 								 dtoh32
 								 (bi_new->
 								  length));
@@ -1595,7 +1592,7 @@ int dhd_iscan_delete_bss(void *dhdp, void *addr, iscan_buf_t *iscan_skip)
 					}
 					break;
 				}
-				bi = (wl_bss_info_t *) ((uintptr) bi +
+				bi = (wl_bss_info_t *)((unsigned long)bi +
 							dtoh32(bi->length));
 			}
 		}
@@ -1650,7 +1647,7 @@ int dhd_iscan_remove_duplicates(void *dhdp, iscan_buf_t *iscan_cur)
 
 		dhd_iscan_delete_bss(dhdp, bi->BSSID.octet, iscan_cur);
 
-		bi = (wl_bss_info_t *) ((uintptr) bi + dtoh32(bi->length));
+		bi = (wl_bss_info_t *)((unsigned long)bi + dtoh32(bi->length));
 	}
 
 done:
